requirements.txt
pandas
numpy
yfinance
mysql-connector-python
beautifulsoup4
requests
streamlit
